requirements.txt
pytest
pytest-logger
pytest-selenium
python-i18n[YAML]

pyvirtualdisplay
